Move most of the quality report code into a module
